# bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Aplicada2Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Reportes": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Reportes"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reportes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Reportes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Reportes"
            }
          }
        }
      }
    },
    "/api/Reportes/{id}": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Reportes"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Reportes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Reportes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Reportes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Solicitudes": {
      "get": {
        "tags": [
          "Solicitudes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Solicitudes"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Solicitudes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Solicitudes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Solicitudes"
            }
          }
        }
      }
    },
    "/api/Solicitudes/{id}": {
      "get": {
        "tags": [
          "Solicitudes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Solicitudes"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Solicitudes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Solicitudes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Solicitudes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Vehiculos": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Vehiculos"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehiculos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Vehiculos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Vehiculos"
            }
          }
        }
      }
    },
    "/api/Vehiculos/{id}": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Vehiculos"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vehiculos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Vehiculos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehiculos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Clientes": {
      "type": "object",
      "properties": {
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "nombres": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "vehiculoId": {
          "format": "int32",
          "type": "integer"
        },
        "vehiculos": {
          "$ref": "#/definitions/Vehiculos"
        }
      },
      "additionalProperties": false
    },
    "Mecanicos": {
      "type": "object",
      "properties": {
        "mecanicoId": {
          "format": "int32",
          "type": "integer"
        },
        "nombres": {
          "type": "string"
        },
        "area": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "disponible": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "Reportes": {
      "type": "object",
      "properties": {
        "reporteId": {
          "format": "int32",
          "type": "integer"
        },
        "concepto": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "mecanicoId": {
          "format": "int32",
          "type": "integer"
        },
        "clientes": {
          "$ref": "#/definitions/Clientes"
        },
        "mecanicos": {
          "$ref": "#/definitions/Mecanicos"
        }
      },
      "additionalProperties": false
    },
    "Solicitudes": {
      "type": "object",
      "properties": {
        "solicitudId": {
          "format": "int32",
          "type": "integer"
        },
        "concepto": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "mecanicoId": {
          "format": "int32",
          "type": "integer"
        },
        "clientes": {
          "$ref": "#/definitions/Clientes"
        },
        "mecanicos": {
          "$ref": "#/definitions/Mecanicos"
        }
      },
      "additionalProperties": false
    },
    "Vehiculos": {
      "type": "object",
      "properties": {
        "vehiculoId": {
          "format": "int32",
          "type": "integer"
        },
        "marca": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "year": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}